+++ cuneiform_src/Kern/hhh/tigerh/h/LAYOUT.H
@@ -212,15 +212,15 @@ void RotatePageToIdeal (void);
 // пересчет координат с учетом наклона
 # define IDEAL_XY(x, y)   \
          {\
-             y = (int16_t) (y - (LONG) x * nIncline / INCLINE_FACTOR);\
-             x = (int16_t) (x + (LONG) y * nIncline / INCLINE_FACTOR);\
+             y = (int16_t) (y - (int32_t) x * nIncline / INCLINE_FACTOR);\
+             x = (int16_t) (x + (int32_t) y * nIncline / INCLINE_FACTOR);\
          }
 
 // возврат к настоящим координатам
 # define REAL_XY(x, y)   \
          {\
-             x = (int16_t) (x - (LONG) y * nIncline / INCLINE_FACTOR);\
-             y = (int16_t) (y + (LONG) x * nIncline / INCLINE_FACTOR);\
+             x = (int16_t) (x - (int32_t) y * nIncline / INCLINE_FACTOR);\
+             y = (int16_t) (y + (int32_t) x * nIncline / INCLINE_FACTOR);\
 		}
 
 /****************************************************************************
