+++ cuneiform_src/Kern/hrstr/match_wd.h
@@ -56,8 +56,8 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 typedef struct tagWeight
 {
-  LONG meas;  //measure
-  LONG nlet;  //letter's numbers
+  int32_t meas;  //measure
+  int32_t nlet;  //letter's numbers
 } Weight;
 
 typedef struct tagMatchWordPar
