+++ cuneiform_src/Kern/hrstr/struct.h
@@ -369,7 +369,7 @@ struct cell_struc {
 	// 112
 	STICK save_baton[MAX_BATON];
 	// 24+112=136
-	LONG dupstart, dupend;
+	int32_t dupstart, dupend;
 #define dup_default     ((cell*)0xffffffff)
 	// 144
 	uchar dlang_dup;
